extract `type_relate` into the inference context as `nll_relate`
The name is not great. Nor is the existence of this code great.  It
should be merged with the main "type relating code" at some point.
